requirements.txt
asgiref==3.9.2
attrs==25.4.0
crispy-bootstrap5==2025.6
crispy-tailwind==1.0.3
Django==5.2.7
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-filter==25.2
django-unfold==0.71.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0